Break up voting tests, add flag to enable skipping bootstrap data
