+++ microservice-vote/pom.xml
@@ -267,7 +267,7 @@
                             </includes>
                             <systemPropertyVariables>
                                 <liberty.test.port>${testServerHttpPort}</liberty.test.port>
-                                <war.name>${warContext}</war.name>
+                                <app.context.root>${warContext}</app.context.root>
                             </systemPropertyVariables>
                         </configuration>
                     </execution>
