+++ microservice-vote/src/main/java/io/microprofile/showcase/vote/api/AttendeeProvider.java
@@ -38,11 +38,6 @@ import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.ext.MessageBodyReader;
 import javax.ws.rs.ext.MessageBodyWriter;
 import javax.ws.rs.ext.Provider;
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 
 import io.microprofile.showcase.vote.model.Attendee;
